requirements.txt
streamlit==1.36.0
crewai==0.16.0
langchain==0.1.8
langchain-openai==0.0.5
langchain-community==0.0.38
langchain-core==0.1.52
pandas==2.2.2
python-dotenv==1.0.1
openai==1.35.14
tiktoken==0.5.2
matplotlib==3.7.1
seaborn==0.12.2